personal/projects/credit-card-optimization/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
jinja2==3.1.6
pydantic==2.10.4
sqlalchemy==2.0.36
pulp==3.3.0
openpyxl==3.1.5
httpx==0.28.1
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pytest==8.3.4
pytest-cov==6.0.0
aiofiles==24.1.0
python-multipart==0.0.20
